refactor(landing): keep WPStatusChangeMetadata in CORE, exported via the status package API
CI (arch_shard_1) caught two boundary contracts the params-object fold crossed:

1. test_integration_boundary: CORE (status/) must not import INTEGRATION
   (sync/). Building the metadata object at the producers had status/emit.py
   importing WPStatusChangeMetadata from sync.events -- allowlist permanently
   closed.
2. test_status_module_boundary: external code must import status symbols via
   `from specify_cli.status import X`, never a status submodule directly.

Fix: move WPStatusChangeMetadata + its from_status_event factory into a new CORE
leaf module status/wp_status_metadata.py, export it from the status package
__init__, and have the external consumers (sync.events, coordination.outbound)
import it as `from specify_cli.status import WPStatusChangeMetadata`. sync.events
re-exports it so `from specify_cli.sync.events import WPStatusChangeMetadata`
keeps working. The typed-param-object design + factory are preserved; only the
layering is corrected. status/emit.py keeps its intra-package import.

No import cycle (status stays importable without sync); ruff + mypy clean;
test_integration_boundary, test_status_module_boundary, test_all_declarations_required,
test_parity (importable-without-sync), and the emit/fanout/coordination suites green.
