+++ src/specify_cli/sync/events.py
@@ -18,9 +18,10 @@ import threading
 import json
 import urllib.parse
 import urllib.request
-from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
+
+from specify_cli.status import WPStatusChangeMetadata
 
 from .feature_flags import is_saas_sync_enabled
 
@@ -291,49 +292,6 @@ def reset_emitter() -> None:
 
 
 # ── Convenience Functions ─────────────────────────────────────
-
-
-@dataclass(frozen=True)
-class WPStatusChangeMetadata:
-    """Optional ``WPStatusChanged`` fields, bundled into one params object.
-
-    S107: ``emit_wp_status_changed`` originally declared these 8 fields as
-    individual keyword parameters, pushing it to 15 total parameters. Every
-    field here mirrors the identically-named keyword ``EventEmitter.
-    emit_wp_status_changed`` (``sync/emitter.py``) already accepts, so
-    threading ``metadata`` through to that call is a plain field-by-field
-    unpack — no behavior change.
-    """
-
-    causation_id: str | None = None
-    policy_metadata: dict[str, Any] | None = None
-    force: bool = False
-    reason: str | None = None
-    review_ref: str | None = None
-    execution_mode: str | None = None
-    evidence: dict[str, Any] | None = None
-    occurred_at: str | None = None
-
-    @classmethod
-    def from_status_event(
-        cls, event: Any, *, policy_metadata: dict[str, Any] | None = None
-    ) -> WPStatusChangeMetadata:
-        """Build from a StatusEvent (the two production fan-out callers).
-
-        ``policy_metadata`` is passed explicitly because callers resolve it
-        separately from the event (status/emit.py resolves its own; the
-        coordination outbound path uses ``event.policy_metadata``).
-        """
-        return cls(
-            causation_id=event.event_id,
-            policy_metadata=policy_metadata,
-            force=event.force,
-            reason=event.reason,
-            review_ref=event.review_ref,
-            execution_mode=event.execution_mode,
-            evidence=event.evidence.to_dict() if event.evidence else None,
-            occurred_at=event.at,
-        )
 
 
 def emit_wp_status_changed(
